% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{mlr3learners.gpfit-package}
\alias{mlr3learners.gpfit}
\alias{mlr3learners.gpfit-package}
\title{mlr3learners.gpfit: Gaussian Process Learner for mlr3}
\description{
Implements Gaussian Process regression using the GPfit package as an mlr3 learner. Provides GP_fit() functionality within the mlr3 framework for machine learning tasks.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/KhalifaSeck/mlr3learners.gpfit}
  \item Report bugs at \url{https://github.com/KhalifaSeck/mlr3learners.gpfit/issues}
}

}
\author{
\strong{Maintainer}: Khalifa SECK \email{seckhalifaa@gmail.com}

}
